% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{cox_step}
\alias{cox_step}
\title{Cox model step-wise fitting}
\usage{
cox_step(x, time = "time", status = "status", surv_form = NULL,
  penalty = "AIC")
}
\arguments{
\item{train}{: train data \cr
time : time variable \cr
status : status variable \cr
method: forward, backward or both \cr
fit : fit model: "Stepwise", "Random forest"}
}
\value{
a coxph forward Stepwise fit object
}
\description{
Fit model by step-wise variable selection.
}
