% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gen_long_dat}
\alias{gen_long_dat}
\title{Generate long data format for stan}
\usage{
gen_long_dat(dat, status = "status", time = "time")
}
\arguments{
\item{d}{a dataset \cr}
}
\value{
d a long dataset
}
\description{
This function generates a long format dataframe for
repeated tte, time dependent covariates.
}
