% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gen_new.frame}
\alias{gen_new.frame}
\title{Expand dataset accordin to calendar time}
\usage{
gen_new.frame(dat, time = "time", id = "patient_id", timepoints)
}
\arguments{
\item{dat}{a data frame \cr}
}
\value{
d a long dataset
}
\description{
Expand dataset accordin to calendar time
}
