% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_stan_data_backup.R
\name{gen_stan_dat_bakckup}
\alias{gen_stan_dat_bakckup}
\title{Generate surv data for stan}
\usage{
gen_stan_dat_bakckup(x, status = "status", time = "time")
}
\arguments{
\item{d}{a dataset \cr}
}
\value{
d a long dataset
}
\description{
This function generates a long format dataframe for
repeated tte, time dependent covariates.
}
