% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gen_stan_surv}
\alias{gen_stan_surv}
\title{Generate data for stan}
\usage{
gen_stan_surv(dat, status = "status", time = "time", timepoints = T)
}
\arguments{
\item{d}{a dataset \cr}
}
\value{
d a long dataset
}
\description{
This function generates a long format dataframe for
repeated tte, time dependent covariates.
}
