% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_surv_bgam.R
\name{get_ci2}
\alias{get_ci2}
\title{Get concordance index}
\usage{
get_ci2(pred_frame, x)
}
\arguments{
\item{x}{data}

\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
mod
}
\description{
Get concordance index from a survbayes model
}
\references{
Terry M. Therneau and Patricia M. Grambsch (2000).
  Modeling Survival Data: Extending the Cox Model.
  Springer, New York. ISBN 0-387-98784-3.
}
\seealso{
[coxph]
}
\author{
Carlos S Traynor
}
\keyword{coxph}
