% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_GAMM.R
\name{get_ibrier_bm}
\alias{get_ibrier_bm}
\title{Model fitting}
\usage{
get_ibrier_bm(x, surv_form, prior = NULL, ...)
}
\arguments{
\item{x}{data}

\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
mod
}
\description{
Fits coxph model.
}
\references{
Terry M. Therneau and Patricia M. Grambsch (2000).
  Modeling Survival Data: Extending the Cox Model.
  Springer, New York. ISBN 0-387-98784-3.
}
\seealso{
[coxph]
}
\author{
Carlos S Traynor
}
\keyword{coxph}
