% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prepare.R
\name{get_plot_new_frame}
\alias{get_plot_new_frame}
\title{Prepare plot with new dataframe}
\usage{
get_plot_new_frame(mod, newdat, timepoints, treatment)
}
\arguments{
\item{post}{posterior draws in matrix format \cr}

\item{strata}{strata}

\item{obs}{original dataframe defaults to ic2surv}
}
\value{
d a plot dataset
}
\description{
This function prepares plot frame
}
