% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ic2surv}
\alias{ic2surv}
\alias{intclust2dat}
\title{ic2surv}
\value{
\item{intclust2dat}{a data frame}
}
\description{
ic2surv
}
\details{
These data are the response variables time/status and includes
 selected variables by glmnet, either using intclust2 only or the pooled dataset and treatment effects.
}
\examples{
data(intclust2dat)
str(intclust2dat)
}
\references{
Pereira, B. et al. Nat. Commun. 7, 2016.
}
\keyword{datasets}
