% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{link.surv}
\alias{link.surv}
\title{GAMMA poisson link}
\usage{
link.surv(post, longdata)
}
\arguments{
\item{post}{posterior draws \cr}

\item{strata}{strata}

\item{obs}{original dataframe defaults to ic2surv}
}
\value{
d a plot dataset
}
\description{
Link model to linear predictor to individual
 survival predictions.
}
