% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_km}
\alias{plot_km}
\title{Kaplan and Meier estimates}
\usage{
plot_km(dat, time = time, status = status, event_type = 1, strata = "1")
}
\arguments{
\item{d}{a dataset \cr
x customise plot title \cr
event_type codification of event type , default DECEASED}
}
\value{
d a clean dataset
}
\description{
This function plots the TKaplan and Meier estimates
}
