% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_GAMM.R
\name{post_surv}
\alias{post_surv}
\title{Posterior Hazard}
\usage{
post_surv(x, surv_form, prior = NULL, iter = 12000, ...)
}
\arguments{
\item{x}{data}

\item{surv_mod}{vector with covariates}
}
\value{
mod
}
\description{
Bayesian semiparametric survival model with GAMM
}
\references{
Terry M. Therneau and Patricia M. Grambsch (2000).
  Modeling Survival Data: Extending the Cox Model.
  Springer, New York. ISBN 0-387-98784-3.
}
\seealso{
[coxph]
}
\author{
Carlos S Traynor
}
\keyword{coxph}
