% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_GAMM.R
\name{pred_sbm}
\alias{pred_sbm}
\title{Prediction of survival from model}
\usage{
pred_sbm(x, surv_form, prior = NULL, ...)
}
\arguments{
\item{x}{data}

\item{mod}{Coxph model object fitted with coxph (survival).}
}
\value{
mod
}
\description{
Fits bayesian semi-parametric survival models
}
\references{
Terry M. Therneau and Patricia M. Grambsch (2000).
  Modeling Survival Data: Extending the Cox Model.
  Springer, New York. ISBN 0-387-98784-3.
}
\seealso{
[coxph]
}
\author{
Carlos S Traynor
}
\keyword{coxph}
