% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{stan_surv_step}
\alias{stan_surv_step}
\title{Bayes model step-wise fitting}
\usage{
stan_surv_step(x, time = "time", status = "status", scope = NULL,
  criterion = "loo", verbose = TRUE, fit = NULL, stan_function = NULL)
}
\arguments{
\item{train}{: train data \cr
time : time variable \cr
status : status variable \cr
method: forward, backward or both \cr
fit : fit model: "Stepwise", "Random forest"}
}
\value{
stan_step fit model
}
\description{
Fit model by step-wise variable selection.
}
