% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{str_strata}
\alias{str_strata}
\title{String method for strata in surv analysis}
\usage{
str_strata(c)
}
\arguments{
\item{c}{a character frame \cr}
}
\value{
d a clean dataset
}
\description{
This function is a string method to facilitate ploting.
}
