% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_GAMM.R
\name{surv_pred_bgam}
\alias{surv_pred_bgam}
\title{Prediction of Survival probability from a Bayesian model}
\usage{
surv_pred_bgam(x, bgam, preProcValues = NULL, ...)
}
\arguments{
\item{x}{data}

\item{bgam}{Bayesian Generalized Additive Models}

\item{...}{Ignored}

\item{preProcess}{Pre-processing information from training set to Test set.}
}
\value{
mod
}
\description{
Predicts the survival probability for each drawn of the posterior distribution.
}
\references{
Wood, S.N. (2006) Generalized Additive Models: An Introduction with R 
 Chapman & Hall/CRC, Boca Raton, Florida. ISBN 1-58488-474-6
}
\seealso{
[gamair]
}
\author{
Carlos S Traynor
}
\keyword{gamair}
