% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_surv.R
\name{pred_surv}
\alias{pred_surv}
\title{Posterior prediction of survival outcome}
\usage{
pred_surv(x = NULL, long_x = NULL, mod, unix = FALSE,
  patient_id = "patient_id")
}
\arguments{
\item{post}{posterior draws in matrix format \cr}

\item{strata}{strata}

\item{obs}{original dataframe defaults to ic2surv}
}
\value{
d a plot dataset
}
\description{
This function predicts the survival from the draws of the posterior.
}
